### 2024-07-07 #6965 @editorialbot
[PRE REVIEW]: pyPLNmodels: A Python package to analyse multivariate high-dimensional count data
**Submitting author:** @bastien-mva (<a href="http://orcid.org/0009-0001-3960-7120">Bastien Batardiere</a>)
**Repository:** https://github.com/PLN-team/pyPLNmodels.git
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.0.78
