+++ paper/paper.md
@@ -34,9 +34,6 @@ patterns and dependency structures. The PLN model is versatile and can be
 extended beyond PCA to accommodate other multivariate statistical tasks, such as Clustering, Linear
 Discriminant Analysis (LDA), and Network inference, described in [@PLNmodels].
 Possible fields of applications include
-<!-- Multivariate abundance data, consisting of observations of multiple interacting -->
-<!-- species from a set of samples, are often collected in ecological studies to -->
-<!-- characterize a community or assemblage of organisms. -->
 
 \begin{itemize}
 \item Ecology: Joint analysis of species abundances is a common task in
@@ -153,92 +150,5 @@ precious advices to build a proper python package.
 # Fundings
 Bastien Bartardière and Julien Chiquet are supported by
 the French ANR grant ANR-18-CE45-0023 Statistics and Machine Learning for Single Cell Genomics (SingleStatOmics).
-<!-- # Mathematics -->
-
-<!-- Single dollars ($) are required for inline mathematics e.g. $f(x) = e^{\pi/x}$ -->
-
-<!-- Double dollars make self-standing equations: -->
-
-<!-- $$\Theta(x) = \left\{\begin{array}{l} -->
-<!-- 0\textrm{ if } x < 0\cr -->
-<!-- 1\textrm{ else} -->
-<!-- \end{array}\right.$$ -->
-
-<!-- You can also use plain \LaTeX for equations -->
-<!-- \begin{equation}\label{eq:fourier} -->
-<!-- \hat f(\omega) = \int_{-\infty}^{\infty} f(x) e^{i\omega x} dx -->
-<!-- \end{equation} -->
-<!-- and refer to \autoref{eq:fourier} from text. -->
-
-<!-- # Citations -->
-
-<!-- Citations to entries in paper.bib should be in -->
-<!-- [rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html) -->
-<!-- format. -->
-
-<!-- If you want to cite a software repository URL (e.g. something on GitHub without a preferred -->
-<!-- citation) then you can do it with the example BibTeX entry below for @fidgit. -->
-
-<!-- For a quick reference, the following citation commands can be used: -->
-<!-- - `@author:2001`  ->  "Author et al. (2001)" -->
-<!-- - `[@author:2001]` -> "(Author et al., 2001)" -->
-<!-- - `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)" -->
-
-<!-- # Figures -->
-
-<!-- Figures can be included like this: -->
-<!-- ![Caption for example figure.\label{fig:example}](figure.png) -->
-<!-- and referenced from text using \autoref{fig:example}. -->
-
-<!-- Figure sizes can be customized by adding an optional second parameter: -->
-<!-- ![Caption for example figure.](figure.png){ width=20% } -->
-
-<!-- # Mathematical details -->
-<!-- Dire dans l'intro le versatile (et laisser l'ouverture pour plus tard) pas -->
-<!-- obligé que c'est implémenté dans Chiquet et al. Nlopt. Essayer de faire rentrer -->
-<!-- les images. -->
-<!-- Numéro de l'ANR. -->
-<!-- We compare pyPLNmodels with or without GPU, PLNmodels and ```GLLVM```. -->
[…]
