### 2024-07-07 #6965 @Bastien-mva
At-mentioning co-authors here: @jchiquet @joon-kwon

Here is a non-exhaustive list of potential reviewers from the JOSS reviewers google sheet (coauthors, please feel free to add more in this thread):
- richford
- ZhenchenHong
- LingfengLuo0510
- mayer79
- mrazomej
- DanWaxman
- simeonhebrew
- matt-graham
- roualdes
- prdm0
- Wrzlprmft
- misken
- vankesteren
- limengbinggz
- larryshamalama
- nickelnine37
- zhenmuw.edu
- janfb
- papachristoumarios
- edrictam
- maximtrp
- Rlucas7
- tylerwmarrs
