### 2024-07-07 8bc9604
add doi for two papers.
