+++ paper/paper.bib
@@ -25,6 +25,7 @@
   title={statsmodels: Econometric and statistical modeling with python},
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
+  doi={10.25080/majora-92bf1922-011},
   year={2010},
 }
 
@@ -63,6 +64,7 @@
       year={2018},
       eprint={1703.06633},
       archivePrefix={arXiv},
+      doi={10.1214/18-aoas1177},
       primaryClass={stat.ME}
 }
 
