### 2024-07-08 #6965 @likeajumprope
> I'm sorry human, I don't understand that. You can see what commands I support by typing:
>
> `@editorialbot commands`

what am I doing wrong haha
