### 2024-07-08 #6965 @crvernon
I believe only editors in chief can run assignments on a name (even if it is your own).  You have to actually run:

`@editorialbot assign me as editor`

No problem though, I've got you on this one.  And thanks!
