### 2024-07-09 #6965 @mrazomej
> Hi @mrazomej 👋 would you be interested in taking this paper on for review?

Yes, I'm interested. Count me in!
