### 2024-07-10 #6969 @Bastien-mva
Figures were not displayed.
