+++ pca.py
@@ -17,7 +17,7 @@ t = time.time()
 plnpca.fit(tol=0.001, verbose=True, nb_max_iteration=700)
 print("time took pln:", (time.time() - t) / 60)
 # plnpca.show()
-fig, axes = plt.subplots(1, 2, figsize=(20, 13), layout="constrained")
+fig, axes = plt.subplots(1, 2, figsize=(20, 10), layout="constrained")
 plnpca.viz(colors=GT, ax=axes[0])
 t = time.time()
 pca = PCA(n_components=2)
