### 2024-07-13 54a08d1
just remove the plot_dict.
