+++ manu.py
@@ -1,22 +1,31 @@
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
+import seaborn as sns
 from pyPLNmodels import Pln, PlnPCA
 import seaborn as sns
 from structpca import PlnPCAsampler, get_linear_params_and_additional_data
+from structpca.pca_sampling_parameters import get_block_components
+import jax.numpy as jnp
+from jax import random
+import jax.numpy.linalg as LA
+from matplotlib.ticker import FormatStrFormatter
 
-
-max_epoch = 400
-max_time = 2
-nb_params = 3
+max_time = 60
+nb_params = 15
 n_samples = 1000
-dim1 = 20
-dim2 = 30
+dim1 = 100
+dim2 = 800
 nb_cov = 2
 latent_dim = 5
-nb_grad_steps = [0, 1, 10, 20]
+nb_grad_steps = ["profiled", 1, 10, 30]
 colors = ["red", "blue", "green", "orange"]
-nb_point = 4
+nb_point = 2000
+
+
+colors = sns.color_palette("viridis")
+my_palette = {"profiled": colors[0], "1": colors[1], "10": colors[2], "30": colors[3]}
+sns.color_palette("viridis")
 
 
 def fit_models(model, seed_param, dim):
@@ -28,8 +37,10 @@ def fit_models(model, seed_param, dim):
         latent_dimension=5,
     )
     if model == Pln:
-        mydict["components"] *= 1
-
+        C = mydict["components"]
+        sigma = C @ (C.T)
+        sigma = C @ (C.T) + jnp.eye(dim)
+        mydict["components"] = LA.cholesky(sigma)
     simulator = PlnPCAsampler.from_dict(mydict)
     counts = simulator.sample(seed=0)
     dict_models = {}
@@ -50,58 +61,66 @@ def fit_models(model, seed_param, dim):
                 add_const=False,
             )
         pln.seed = seed_param
-        if nb_grad_step == 0:
-            pln.fit(max_time=max_time)
+        if nb_grad_step == "profiled":
+            if model == Pln:
+                pln.fit(max_time=max_time)
+            else:
+                pln = None
         else:
-            pln.fit_vem(nb_gradient_steps=nb_grad_step, max_time=max_time)
+            if model == Pln:
+                pln.fit_vem(nb_gradient_steps=nb_grad_step, max_time=max_time)
+            else:
+                pln.fit_vem(nb_gradient_steps=nb_grad_step, max_time=2 * max_time)
         dict_models[nb_grad_step] = pln
     return dict_models
 
 
-fig, all_axes = plt.subplots(2, 2, figsize=(20, 10), layout="constrained")
+# def plot_dict(dict_model, modeltype, ax, tolegend):
+#     for i in range(len(nb_grad_steps)):
+#         nb_grad_step = nb_grad_steps[i]
+#         model = dict_model[nb_grad_step]
+#         if tolegend is True:
+#             ax.plot(
+#                 model.running_times,
+#                 -np.array(model._elbos_list),
+#                 label=nb_grad_step,
+#                 color=colors[i],
+#             )
+#         else:
+#             ax.plot(model.running_times, -np.array(model._elbos_list), color=colors[i])
 
 
-def plot_dict(dict_model, modeltype, ax, tolegend):
-    for i in range(len(nb_grad_steps)):
-        nb_grad_step = nb_grad_steps[i]
-        model = dict_model[nb_grad_step]
-        if tolegend is True:
-            ax.plot(
-                model.running_times,
[…]
