+++ manu.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
-from pyPLNmodels import Pln, PlnPCA
+from pyPLNmodels import PlnPCA
 import seaborn as sns
 from structpca import PlnPCAsampler, get_linear_params_and_additional_data
 from structpca.pca_sampling_parameters import get_block_components
@@ -11,20 +11,18 @@ from jax import random
 import jax.numpy.linalg as LA
 from matplotlib.ticker import FormatStrFormatter
 
-max_time = 60
-nb_params = 15
-n_samples = 1000
+max_time = 5
+nb_params = 3
+n_samples = 200
 dim1 = 100
-dim2 = 800
+dim2 = 200
 nb_cov = 2
 latent_dim = 5
-nb_grad_steps = ["profiled", 1, 10, 30]
+nb_grad_steps = ["Singular", 10, 30]
 colors = ["red", "blue", "green", "orange"]
-nb_point = 2000
+nb_point = 200
 
 
-colors = sns.color_palette("viridis")
-my_palette = {"profiled": colors[0], "1": colors[1], "10": colors[2], "30": colors[3]}
 sns.color_palette("viridis")
 
 
@@ -34,60 +32,27 @@ def fit_models(model, seed_param, dim):
         n_samples=n_samples,
         nb_cov=nb_cov,
         dim=dim,
-        latent_dimension=5,
+        latent_dimension=latent_dim,
     )
-    if model == Pln:
-        C = mydict["components"]
-        sigma = C @ (C.T)
-        sigma = C @ (C.T) + jnp.eye(dim)
-        mydict["components"] = LA.cholesky(sigma)
+    mydict["exog"] = mydict["exog"].at[:, 0].set(1)
     simulator = PlnPCAsampler.from_dict(mydict)
     counts = simulator.sample(seed=0)
     dict_models = {}
     for nb_grad_step in nb_grad_steps:
-        if model == PlnPCA:
-            pln = model(
-                np.asarray(counts),
-                exog=np.asarray(simulator.exog),
-                offsets=np.asarray(simulator.offsets),
-                rank=latent_dim,
-                add_const=False,
-            )
-        else:
-            pln = model(
-                np.asarray(counts),
-                exog=np.asarray(simulator.exog),
-                offsets=np.asarray(simulator.offsets),
-                add_const=False,
-            )
+        pln = model(
+            np.asarray(counts),
+            exog=np.asarray(simulator.exog),
+            offsets=np.asarray(simulator.offsets),
+            rank=latent_dim,
+            add_const=False,
+        )
         pln.seed = seed_param
-        if nb_grad_step == "profiled":
-            if model == Pln:
-                pln.fit(max_time=max_time)
-            else:
-                pln = None
+        if nb_grad_step == "Singular":
+            pln.fit(max_time=max_time)
         else:
-            if model == Pln:
-                pln.fit_vem(nb_gradient_steps=nb_grad_step, max_time=max_time)
-            else:
-                pln.fit_vem(nb_gradient_steps=nb_grad_step, max_time=2 * max_time)
[…]
