+++ parametrizations.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
-from pyPLNmodels import Pln, PlnPCA, Pln0
+from pyPLNmodels import Pln, PlnPCA, Pln0, load_scrna
 import seaborn as sns
 from structpca import PlnPCAsampler, get_linear_params_and_additional_data
 from structpca.pca_sampling_parameters import get_block_components
@@ -11,18 +11,20 @@ from jax import random
 import jax.numpy.linalg as LA
 from matplotlib.ticker import FormatStrFormatter
 
-max_time = 60
-nb_params = 5
-n_samples = 1000
+max_time = 2
+nb_params = 4
+n_samples = 400
 dim1 = 100
-dim2 = 800
+dim2 = 200
 nb_cov = 2
 parametrizations = ["Profiled", "uPLN-0", "uPLN-PF"]
 colors = ["red", "blue", "green", "orange"]
-nb_point = 15
+nb_point = 100
 
 
 def fit_models(seed_param, dim):
+    data = load_scrna(n_samples=n_samples, dim=dim)
+    n_samples_ = data["endog"].shape[0]
     mydict = get_linear_params_and_additional_data(
         seed=seed_param,
         n_samples=n_samples,
@@ -38,31 +40,37 @@ def fit_models(seed_param, dim):
     simulator = PlnPCAsampler.from_dict(mydict)
     counts = simulator.sample(seed=0)
     dict_models = {}
+    counts = data["endog"]
+    exog = np.ones((n_samples_, 1))
+    offsets = np.zeros((n_samples_, dim))
+    print("offsets", offsets.shape)
     for parametrization in parametrizations:
         if parametrization == "0":
             pln = Pln0(
                 np.asarray(counts),
-                exog=np.asarray(simulator.exog),
-                offsets=np.asarray(simulator.offsets),
+                exog=np.asarray(exog),
+                offsets=np.asarray(offsets),
                 add_const=False,
             )
         elif parametrization == "PF":
             pln = PlnPCA(
                 np.asarray(counts),
-                exog=np.asarray(simulator.exog),
-                offsets=np.asarray(simulator.offsets),
+                exog=np.asarray(exog),
+                offsets=np.asarray(offsets),
                 rank=dim,
                 add_const=False,
             )
         else:
             pln = Pln(
                 np.asarray(counts),
-                exog=np.asarray(simulator.exog),
-                offsets=np.asarray(simulator.offsets),
+                exog=np.asarray(exog),
+                offsets=np.asarray(offsets),
                 add_const=False,
             )
         pln.seed = seed_param
-        pln.fit(max_time=max_time)
+        pln.fit(
+            max_time=max_time,
+        )
         dict_models[parametrization] = pln
     return dict_models
 
@@ -80,7 +88,7 @@ def append_dict(dict_model):
             np.linspace(0, len(model.running_times) - 1, nb_point)
         ):
             i = int(i)
-            df = df.append(
+            dict_new = (
                 {
                     "model": model.__class__.__name__,
                     "dim": model.dim,
@@ -90,8 +98,9 @@ def append_dict(dict_model):
                     "time": model.running_times[i],
                     "absc": time_number,
                 },
-                ignore_index=True,
             )
+            new_df = pd.DataFrame.from_dict(dict_new)
+            df = pd.concat([df, new_df], ignore_index=True)
 
 
[…]
