### 2024-08-01 b270e35
add svd solver and dict_plots for plots.
