### 2024-08-05 daf7a2c
add markers in the viz function.
