### 2024-08-05 ce42234
nicer figure (colors).
