### 2024-08-05 71f43f0
add nicer color for the benchmark plot.
