### 2024-08-05 #6969 @Bastien-mva
nicer color in figure.
