### 2024-08-05 9df65b8
right figure with right color in the .md.
