### 2024-08-05 #6969 @mrazomej
First, I must apologize for the delayed comments on this submission.

There are two main comments I would like to make about the state of the software and the paper.

1. The [software documentation](https://bbatardiere.pages.mia.inra.fr/pyplnmodels/) needs significant work. As it stands, no user without previous experience with the methods offered by the package--and I would say even with previous knowledge of the methods--could get started with the information provided. There are no instructions on how to get started (a detailed `quickstart` section would be greatly appreciated) or the differences between the different models. The documentation only contains the automatically generated API information from the function docstrings. Even the [`quickstart` section](https://github.com/PLN-team/pyPLNmodels/blob/main/README.md) on the `README.md` file does not explain anything being done.
1.a. On that note, between the paper and the documentation, it was not clear to me what the software's capabilities are. I felt that without reading the [previous publication](https://projecteuclid.org/journals/annals-of-applied-statistics/volume-12/issue-4/Variational-inference-for-probabilistic-Poisson-PCA/10.1214/18-AOAS1177.full) I stand no chance of understanding what your package can do, even though I have enough background knowledge on related areas. As the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) indicate:

> Documentation
```
There should be sufficient documentation for you, the reviewer, to understand the core functionality of the software under review. A README file (or equivalent) should include a high-level overview of this documentation.
```

2. The paper also needs significant work. For example, the `Summary` section does not feel like an actual summary of the functionality of the software. @likeajumprope can correct me if I'm wrong, but I think of this section as the **Abstract** of the paper, where one can quickly assess whether the paper is relevant or not for one's intentions. Even the [JOSS guidelines](https://joss.readthedocs.io/en/latest/paper.html) indicate that the paper should include:
```
A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
```
2.a Fig. 2 does not have a caption that explains, among other things, what the legend labels are. Furthermore, when doing this kind of comparison, it would be handy to highlight which of the curves are directly related to your software. Something that can be added in a detailed caption or in the legend, something like `model name (ours)` to guide the reader.

These are my first two initial comments. These changes are necessary for me to properly assess the software and the entire submission, as I need the most basic tools to understand the software's core functionality. From what I can get from glancing at the previous publications, this is an exciting piece of software. But my understanding is that, as a `JOSS` publication, it needs to exist as a stand-alone piece of work where users (and reviewers) can utilize the software solely from the paper and the documentation information.
