### 2024-08-05 a0a5d57
add lower tolerance for simulations: 0.1 > 0.2.
