+++ docs/source/conf.py
@@ -35,7 +35,9 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.napoleon",
     "sphinx_rtd_theme",
+    "myst_parser",
 ]
+source_suffix = [".rst", ".md"]
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
 
