### 2024-08-05 ec71ec2
tries math in readme.
