### 2024-08-06 8220e24
take_log_offsets was not available in PlnPCA.
