+++ pyPLNmodels/models.py
@@ -457,7 +457,6 @@ class _model(ABC):
             criterion = self._update_criterion_args(loss)
             if abs(criterion) < tol and self._onlyonebatch is True:
                 stop_condition = True
-            print("elbo", loss)
             if nb_epoch_done % 25 == 0:
                 for name_param, param in self.model_parameters.items():
                     mse_param = 0 if param is None else mse(param).detach().item()
