### 2024-08-06 b082993
log scale when showing the matrix.
