### 2024-08-06 bdf20b4
right visualization of the positions.
