+++ README.md
@@ -18,22 +18,33 @@
 <!-- > comprehensive introduction. -->
 
 ##  Getting started
-The getting started can be found [here](Getting_started.ipynb). If you need just a quick view of the package, see the quickstart next.
+A notebook to get started can be found
+[HERE](https://github.com/PLN-team/pyPLNmodels/blob/main/Getting_started.ipynb).
+If you need just a quick view of the package, see the quickstart next.
 
 ## 🛠 Installation
 
 **pyPLNmodels** is available on
 [pypi](https://pypi.org/project/pyPLNmodels/). The development
-version is available on [GitHub](https://github.com/PLN-team/pyPLNmodels).
+version is available on [GitHub](https://github.com/PLN-team/pyPLNmodels) and [GitLab](https://gitlab.com/Bastien-mva/pyplnmodels).
 
 ### Package installation
-
 ```
 pip install pyPLNmodels
 ```
 
+
 ## Statistical description
-Consider $\mathbf Y$ a count matrix consisting of $n$ rows and $p$ columns (denoted as ```endog``` in the package).
+
+For those unfamiliar with the concepts of Poisson or Gaussian random variables,
+it is not necessary to delve into these statistical descriptions. The key
+takeaway is as follows:
+This package is designed to analyze multi-dimensional count data. It
+effectively extracts significant information, such as
+the mean, the relationships with covariates, and the correlation between count
+variables, in a manner appropriate for count data.
+
+Consider $\mathbf Y$ a count matrix (denoted as ```endog``` in the package) consisting of $n$ rows and $p$ columns.
 It is assumed that each individual $\mathbf Y_i$, that is the $i^{\text{th}}$
 row of $\mathbf Y$, is independent from the others and follows a Poisson
 lognormal distribution:
@@ -62,40 +73,46 @@ oaks = load_oaks()
 
 ### How to specify a model
 Each model can be specified in two distinct manners:
-- by formula (similar to R), where a data frame is passed and the formula is specified using the  ```from_formula``` initialization:
-```model = Model.from_formula("endog ~ 1  + covariate_name ", data = oaks)```
+
+* by formula (similar to R), where a data frame is passed and the formula is specified using the  ```from_formula``` initialization:
+
+```model = Model.from_formula("endog ~ 1  + covariate_name ", data = oaks)# not run```
+
 We rely to the [patsy](https://github.com/pydata/patsy) package for the formula parsing.
-- by specifying the endog, exog, and offsets matrices directly:
-```model = Model(endog = oaks["endog"], exog = oaks[["covariate_name"]], offsets = oaks[["offset_name"]])```
+
+* by specifying the endog, exog, and offsets matrices directly:
+
+```model = Model(endog = oaks["endog"], exog = oaks[["covariate_name"]], offsets = oaks[["offset_name"]])# not run```
 
 The parameters `exog` and `offsets` are optional. By default,
 `exog` is set to represent an intercept, which is a vector of ones. Similarly,
 `offsets` defaults to a matrix of zeros.
 
+### Unpenalized Poisson lognormal model (aka `Pln`)
 
-### Unpenalized Poisson lognormal model (aka PLN)
-
-This is the building-block of the package. It fits a Poisson lognormal model to the data.
+This is the building-block of the models implemented in this package. It fits a Poisson lognormal model to the data:
 ```
-pln = Pln.from_formula("endog ~ 1  + tree ", data = oaks, take_log_offsets = True)
+pln = Pln.from_formula("endog ~ 1  + tree ", data = oaks)
 pln.fit()
 print(pln)
[…]
