+++ docs/source/readme.rst
@@ -1,0 +1,222 @@
+PLNmodels: Poisson lognormal models
+===================================
+
+   The Poisson lognormal model and variants can be used for analysis of
+   mutivariate count data. This package implements efficient algorithms
+   extracting meaningful data from difficult to interpret and complex
+   multivariate count data. It has been built to scale on large datasets
+   even though it has memory limitations. Possible fields of
+   applications include - Genomics (number of times a gene is expressed
+   in a cell) - Ecology (species abundances)
+
+   One main functionality is to normalize the count data to obtain more
+   valuable data. It also analyse the significance of each variable and
+   their correlation as well as the weight of covariates (if available).
+
+Getting started
+---------------
+
+A notebook to get started can be found
+`HERE <https://github.com/PLN-team/pyPLNmodels/blob/main/Getting_started.ipynb>`__.
+If you need just a quick view of the package, see the quickstart next.
+
+🛠 Installation
+--------------
+
+**pyPLNmodels** is available on
+`pypi <https://pypi.org/project/pyPLNmodels/>`__. The development
+version is available on
+`GitHub <https://github.com/PLN-team/pyPLNmodels>`__ and
+`GitLab <https://gitlab.com/Bastien-mva/pyplnmodels>`__.
+
+Package installation
+~~~~~~~~~~~~~~~~~~~~
+
+::
+
+   pip install pyPLNmodels
+
+Statistical description
+-----------------------
+
+For those unfamiliar with the concepts of Poisson or Gaussian random
+variables, it is not necessary to delve into these statistical
+descriptions. The key takeaway is as follows: This package is designed
+to analyze multi-dimensional count data. It effectively extracts
+significant information, such as the mean, the relationships with
+covariates, and the correlation between count variables, in a manner
+appropriate for count data.
+
+Consider :math:`\mathbf Y` a count matrix (denoted as ``endog`` in the
+package) consisting of :math:`n` rows and :math:`p` columns. It is
+assumed that each individual :math:`\mathbf Y_i`, that is the
+:math:`i^{\text{th}}` row of :math:`\mathbf Y`, is independent from the
+others and follows a Poisson lognormal distribution:
+
+.. math::
+
+   \mathbf Y_{i}\sim \mathcal P(\exp(\mathbf Z_{i})), \quad \mathbf Z_i \sim
+   \mathcal N(\mathbf o_i + \mathbf B ^{\top} \mathbf x_i, \mathbf \Sigma),
+
+where :math:`\mathbf x_i \in \mathbb R^d` (``exog``) and
+:math:`\mathbf o_i \in \mathbb R^p` (``offsets``) are user-specified
+covariates and offsets. The matrix :math:`\mathbf B` is a
+:math:`d\times p` matrix of regression coefficients and
+:math:`\mathbf \Sigma` is a :math:`p\times p` covariance matrix. The
+goal is to estimate the parameters :math:`\mathbf B` and
+:math:`\mathbf \Sigma`, denoted as ``coef`` and ``covariance`` in the
+package, respectively. A normalization procedure adequate to count data
+can be applied by extracting the ``latent_variables``
+:math:`\mathbf Z_i` once the parameters are learned.
+
+⚡️ Quickstart
+-------------
+
+The package comes with an ecological data set to present the
+functionality:
+
+::
+
+   import pyPLNmodels
+   from pyPLNmodels.models import PlnPCAcollection, Pln, ZIPln
[…]
