### 2024-08-07 57805a7
latent_positions is better than latent_position
