+++ docs/source/documentation.rst
@@ -1,0 +1,31 @@
+.. _documentation-label:
+API documentation
+=================
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Classes
+
+   ./plnpcacollection.rst
+   ./plnpca.rst
+   ./pln.rst
+   ./zipln.rst
+
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Load saved models
+
+   ./load.rst
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Sampling PLN data
+
+   ./sampling.rst
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
