+++ README.md
@@ -18,23 +18,52 @@
 <!-- > comprehensive introduction. -->
 
 ##  Getting started
-The getting started can be found [here](Getting_started.ipynb). If you need just a quick view of the package, see the quickstart next.
+A notebook to get started can be found
+[HERE](https://github.com/PLN-team/pyPLNmodels/blob/main/Getting_started.ipynb).
+If you need just a quick view of the package, see the quickstart next.
 
 ## 🛠 Installation
 
 **pyPLNmodels** is available on
 [pypi](https://pypi.org/project/pyPLNmodels/). The development
-version is available on [GitHub](https://github.com/PLN-team/pyPLNmodels).
+version is available on [GitHub](https://github.com/PLN-team/pyPLNmodels) and [GitLab](https://gitlab.com/Bastien-mva/pyplnmodels).
 
 ### Package installation
-
 ```
 pip install pyPLNmodels
 ```
 
+
+## Statistical description
+
+For those unfamiliar with the concepts of Poisson or Gaussian random variables,
+it is not necessary to delve into these statistical descriptions. The key
+takeaway is as follows:
+This package is designed to analyze multi-dimensional count data. It
+effectively extracts significant information, such as
+the mean, the relationships with covariates, and the correlation between count
+variables, in a manner appropriate for count data.
+
+Consider $\mathbf Y$ a count matrix (denoted as ```endog``` in the package) consisting of $n$ rows and $p$ columns.
+It is assumed that each individual $\mathbf Y_i$, that is the $i^{\text{th}}$
+row of $\mathbf Y$, is independent from the others and follows a Poisson
+lognormal distribution:
+$$\mathbf Y_{i}\sim \mathcal P(\exp(\mathbf Z_{i})), \quad \mathbf Z_i \sim
+\mathcal N(\mathbf o_i + \mathbf B ^{\top} \mathbf x_i, \mathbf \Sigma),$$
+where $\mathbf x_i \in \mathbb R^d$ (`exog`) and $\mathbf o_i \in \mathbb R^p$ (`offsets`) are
+user-specified covariates and offsets. The matrix $\mathbf B$ is a $d\times p$
+matrix of regression coefficients and $\mathbf \Sigma$ is a $p\times p$
+covariance matrix. The goal is to estimate the parameters $\mathbf B$ and
+$\mathbf \Sigma$, denoted as ```coef``` and ```covariance``` in the package,
+respectively. A normalization procedure adequate to count data can be applied
+by extracting the ```latent_variables``` $\mathbf Z_i$ once the parameters are learned.
+
+
+
+
 ## ⚡️ Quickstart
 
-The package comes with an ecological data set to present the functionality
+The package comes with an ecological data set to present the functionality:
 ```
 import pyPLNmodels
 from pyPLNmodels.models import PlnPCAcollection, Pln, ZIPln
@@ -42,41 +71,122 @@ from pyPLNmodels.oaks import load_oaks
 oaks = load_oaks()
 ```
 
-### Unpenalized Poisson lognormal model (aka PLN)
+### How to specify a model
+Each model can be specified in two distinct manners:
 
+* by formula (similar to R), where a data frame is passed and the formula is specified using the  ```from_formula``` initialization:
+
+```model = Model.from_formula("endog ~ 1  + covariate_name ", data = oaks)# not run```
+
+We rely to the [patsy](https://github.com/pydata/patsy) package for the formula parsing.
+
+* by specifying the endog, exog, and offsets matrices directly:
+
+```model = Model(endog = oaks["endog"], exog = oaks[["covariate_name"]], offsets = oaks[["offset_name"]])# not run```
+
[…]
