### 2024-08-07 f1c0df5
add pandoc image and run pandoc before building the doc.
