+++ .gitlab-ci.yml
@@ -61,12 +61,16 @@ publish_package:
 
 pages:
   stage: publish
+  image:
+    name: pandoc
+    entrypoint: ["/bin/sh", "-c"]
+  script:
+    pandoc README.md --from markdown --to rst -s -o docs/source/readme.rst
   image: "registry.forgemia.inra.fr/bbatardiere/docker-image-pandas-torch-sphinx-jupyter"
   before_script:
     - pip install '.[build-doc]'
   script:
     - pip install .
-    - pandoc README.md --from markdown --to rst -s -o docs/source/readme.rst
     - make -C docs html
     - rm -rf public/
     - cp -r docs/build/html public
