+++ .gitlab-ci.yml
@@ -65,7 +65,7 @@ pages:
     name: pandoc
     entrypoint: ["/bin/sh", "-c"]
   script:
-    pandoc README.md --from markdown --to rst -s -o docs/source/readme.rst
+     - pandoc README.md --from markdown --to rst -s -o docs/source/readme.rst
   image: "registry.forgemia.inra.fr/bbatardiere/docker-image-pandas-torch-sphinx-jupyter"
   before_script:
     - pip install '.[build-doc]'
