### 2024-08-07 f3194fb
install pandoc before running scripts.
