+++ .gitlab-ci.yml
@@ -60,6 +60,7 @@ publish_package:
     - tags
 
 pages:
+  stage: publish
   image: "registry.forgemia.inra.fr/bbatardiere/docker-image-pandas-torch-sphinx-jupyter"
   before_script:
     - pip install '.[build-doc]'
