### 2024-08-07 b198778
install pandoc with dpkg and wget.
