### 2024-08-07 5a4a2a5
run the tests for the dev branch on the ci.
