+++ .gitlab-ci.yml
@@ -29,7 +29,7 @@ tests:
     - pytest .
   only:
     - main
-    # - dev
+    - dev
     - tags
 
 
@@ -79,4 +79,3 @@ pages:
     - docker
   only:
     - tags
-    - dev
