### 2024-08-07 0c411c1
does not show the coef if coef is None.
