+++ viz.r
@@ -18,15 +18,14 @@ file_pln_r = "df_pln_r.csv"
 file_plnpca_r = "df_plnpca_r.csv"
 file_gllvm = "df_gllvm.csv"
 filenames = c(file_pln_python, file_plnpca_python, file_pln_python_GPU, file_plnpca_python_GPU, file_gllvm, file_pln_r, file_plnpca_r)
-modelnames = c("py-PLN-CPU", "py-PLN-PCA-CPU", "py-PLN-GPU", "py-PLN-PCA-GPU", "GLLVM", "R-PLN", "R-PLN-PCA")
+modelnames = c("py-PLN-CPU (ours)", "py-PLN-PCA-CPU (ours)", "py-PLN-GPU (ours)", "py-PLN-PCA-GPU (ours)", "GLLVM", "R-PLN", "R-PLN-PCA")
 ###            GLLVM          pyplncpu         pyplngpu       pyplnpcaCPU      pyPLNPCAGPU  RPLN    RPLNPCA
 
 
 col_viridis = viridis(10)
 
-# colors =     c("black",        "blue",          "blue",      "red",        "red",    "cornflowerblue", "firebrick")
-colors =     c(col_viridis[[1]],    col_viridis[[4]],          col_viridis[[4]],      col_viridis[[8]],        col_viridis[[8]],    col_viridis[[3]], col_viridis[[9]])
-linestyles = c("solid",        "solid",        "dashed",     "solid",      "dashed",   "solid",   "solid")
+colors =     c(col_viridis[[4]],    col_viridis[[8]],          col_viridis[[4]],      col_viridis[[8]],        col_viridis[[1]],    col_viridis[[3]], col_viridis[[9]])
+linestyles = c("solid",        "solid",        "dashed",     "dashed",      "solid",   "solid",   "solid")
 # filenames = c(file_pln_r,file_plnpca_r, file_gllvm)
 # modelnames = c("R-Pln", "R-PlnPCA", "GLLVM")
 
@@ -62,8 +61,8 @@ get_plot_i <- function(i){
     current_plot <- ggplot(df) + geom_line(aes(x =dim, y = df[,colnames(df)[i]], group = Model, col = Model, linetype = Model))
     current_plot <- current_plot + ggtitle(paste("n = ", colnames(df)[i])) + labs(y = "Running time (seconds)", x = TeX("Number of variables $p$"))
     current_plot <- current_plot + scale_y_log10()
-    current_plot <- current_plot + theme_bw()+ theme(plot.title = element_text(hjust = 0.5),legend.text = element_text(size=7)) + guides(col = guide_legend(nrow = 1, byrow = FALSE, title = ""), linetype =  guide_legend(nrow = 1, byrow = FALSE, title = "") )
-    current_plot <- current_plot + scale_color_manual(values=colors) + scale_linetype_manual(values=linestyles)
+    current_plot <- current_plot + scale_color_manual(values=colors, limits = modelnames) + scale_linetype_manual(values=linestyles, limits = modelnames)
+    current_plot <- current_plot + theme_bw()+ theme(plot.title = element_text(hjust = 0.5),legend.text = element_text(size=7)) + guides(col = guide_legend(nrow = 2, byrow = TRUE, title = ""), linetype =  guide_legend(nrow = 1, byrow = FALSE, title = "") )
 }
 
 pdf("paper/figures/plots_benchmark.pdf")
