### 2024-08-07 7576c32
change the figure and the caption.
