### 2024-08-07 34943a2
continue writing.
