+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'pyPLNmodels: A Python package to analyse multivariate high-dimensional count data'
+title: 'pyPLNmodels: A Python package to analyze multivariate high-dimensional count data'
 tags:
   - Python
   - count data
@@ -24,35 +24,56 @@ bibliography: paper.bib
 ---
 
 # Summary
-High dimensional count data are hard to analyse as is, and normalization must
-be performed but standard normalization does not fit to the characteristics of
+High dimensional count data are complex to analyze as is, and normalization must
+be performed, but standard normalization does not fit the characteristics of
 count data. The Poisson LogNormal(PLN)  [@PLN] and its Principal Component
 Analysis variant PLN-PCA [@PLNPCA] are two-sided latent variable models allowing both
-suitable normalization and analysis of multivariate count data. The latent
-variable is structured as a multivariate Gaussian variable, modelling complex
-patterns and dependency structures. The PLN model is versatile and can be
-extended beyond PCA to accommodate other multivariate statistical tasks, such as Clustering, Linear
-Discriminant Analysis (LDA), and Network inference, described in [@PLNmodels].
-Possible fields of applications include
+suitable normalization and analysis of multivariate count data, implemented in this package.
 
+Consider $\mathbf Y$ a count matrix consisting of $n$ rows and $p$ columns.
+It is assumed that each individual $\mathbf Y_i$, that is the $i^{\text{th}}$
+row of $\mathbf Y$, is independent of the others and follows a Poisson
+lognormal distribution:
+$$\mathbf Y_{i}\sim \mathcal P(\exp(\mathbf Z_{i})), \quad \mathbf Z_i \sim
+\mathcal N(\mathbf o_i + \mathbf B ^{\top} \mathbf x_i, \boldsymbol{\Sigma}),$$
+where $\mathbf x_i \in \mathbb R^d$ and $\mathbf o_i \in \mathbb R^p$ are
+user-specified covariates and offsets (with default values if not available). The matrix $\mathbf B$ is a $d\times p$
+matrix of regression coefficients and $\boldsymbol{\Sigma}$ is a $p\times p$ covariance matrix. The variables $\mathbf Z_i$, known as *latent variables*,
+are not directly observable. However, from a statistical perspective,
+they provide more informative insights compared to the observed variables
+$\mathbf Y_i$. The unknown parameters $\mathbf B$ and
+$\boldsymbol{\Sigma}$ facilitates the analysis of
+dependencies between variables and
+the impact of covariates. The primary objective of the package is to estimate these
+parameters and retrieve the latent variables $\mathbf Z_i$.  Extracting
+those latent variables may serve as a normalization procedure adequate to count data.
+
+The only difference between the PLN and PLN-PCA models is that the latter
+assumes a low-rank structure on the covariance matrix, which is helpful for
+dimension
+reduction. Other variants of the PLN model exist, which are detailed in
+the work of [@PLNmodels].
+
+# Fields of applications and functionalities
+Possible fields of applications include
 \begin{itemize}
 \item Ecology: Joint analysis of species abundances is a common task in
 ecology, whose goal is to understand the interaction between species to
 characterize a community, given a matrix of abundances in different sites with abundances given by
 $$Y_{ij} = \text{number of species } j \text{ observed in site } i .$$
-Specifically, it aims to establish potential dependencies, competitive interactions, and predatory dynamics.
+<!-- Specifically, it aims to establish potential dependencies, competitive interactions, and predatory dynamics. -->
 Additionally, the PLN models seek to explain the impact of covariates (when available), such as temperature, altitude, and other
-  relevant factors, on the observed abundances.
-\item Genomics: High throughput sequencing technologies now allow to quantify at the level of
[…]
