+++ viz.r
@@ -24,8 +24,8 @@ modelnames = c("py-PLN-CPU (ours)", "py-PLN-PCA-CPU (ours)", "py-PLN-GPU (ours)"
 
 col_viridis = viridis(10)
 
-colors =     c(col_viridis[[4]],    col_viridis[[8]],          col_viridis[[4]],      col_viridis[[8]],        col_viridis[[1]],    col_viridis[[3]], col_viridis[[9]])
-linestyles = c("solid",        "solid",        "dashed",     "dashed",      "solid",   "solid",   "solid")
+colors =     c(col_viridis[[5]],    col_viridis[[8]],          col_viridis[[5]],      col_viridis[[8]],        col_viridis[[1]],    col_viridis[[5]], col_viridis[[8]])
+linestyles = c("solid",        "solid",        "dashed",     "dashed",      "dotted",   "dotted",   "dotted")
 # filenames = c(file_pln_r,file_plnpca_r, file_gllvm)
 # modelnames = c("R-Pln", "R-PlnPCA", "GLLVM")
 
@@ -56,11 +56,12 @@ remove_legend <- function(myplot){
 
 
 get_plot_i <- function(i){
-    current_plot <- ggplot(df) + geom_line(aes(x =dim, y = df[,colnames(df)[i]], group = Model, col = Model, linetype = Model))
+    current_plot <- ggplot(df) + geom_line(aes(x =dim, y = df[,colnames(df)[i]], group = Model, col = Model, linetype = Model), shape = 23, size = 0.75)
     current_plot <- current_plot + ggtitle(paste("n = ", colnames(df)[i])) + labs(y = "Running time (seconds)", x = TeX("Number of variables $p$"))
     current_plot <- current_plot + scale_y_log10()
     current_plot <- current_plot + scale_color_manual(values=colors, limits = modelnames) + scale_linetype_manual(values=linestyles, limits = modelnames)
-    current_plot <- current_plot + theme_bw()+ theme(plot.title = element_text(hjust = 0.5),legend.text = element_text(size=7)) + guides(col = guide_legend(nrow = 2, byrow = TRUE, title = ""), linetype =  guide_legend(nrow = 1, byrow = FALSE, title = "") )
+    current_plot <- current_plot + theme_bw()+ theme(legend.key.width = unit(1.2,"cm"), plot.title = element_text(hjust = 0.5),legend.text = element_text(size=7)) + guides(col = guide_legend(nrow = 2, byrow = TRUE, title = ""), linetype =  guide_legend(nrow = 1, byrow = FALSE, title = ""), linetype = guide_legend(override.aes = list(size = 4)) )
+    # current_plot <- current_plot + guides()
 }
 
 pdf("paper/figures/plots_benchmark.pdf")
