### 2024-08-07 c012871
better colors and linetypes.
