### 2024-08-08 4926c96
add tests and consistency.
