+++ sandwich.py
@@ -1,0 +1,185 @@
+import torch
+from pyPLNmodels import sample_pln, Pln, get_simulation_parameters
+import math
+from tqdm import tqdm
+import seaborn as sns
+import matplotlib.pyplot as plt
+from itertools import product
+import pandas as pd
+import numpy as np
+
+nb_max_iter = 700
+nb_seed = 3
+ns = [250, 500, 1000]
+dims = [20, 40, 50]
+nb_covs = [1, 2, 3]
+
+
+def normalizing(Theta, true_Theta, hess, n):
+    ## Studentization of the error
+    vec_Theta = Theta.flatten()
+    vec_true_Theta = true_Theta.flatten()
+    C_hess = torch.linalg.cholesky(hess)
+    n01 = math.sqrt(n) * torch.matmul(C_hess, vec_Theta - vec_true_Theta)
+    return n01
+
+
+class Fisher_Pln:
+    def __init__(self, Y, A, X, d, p, n, S, Omega, Sigma):
+        self.Y = Y
+        self.A = A
+        self.X = X
+        self.d = d
+        self.p = p
+        self.n = n
+        self.S = S
+        self.Omega = Omega
+        self.Sigma = Sigma
+
+    def getDnTheta(self):
+        YmA = self.Y - self.A
+        YmA_outer = torch.matmul(YmA.unsqueeze(2), YmA.unsqueeze(1))
+        xxt_outer = torch.matmul(self.X.unsqueeze(2), self.X.unsqueeze(1))
+        res = torch.zeros(self.d * self.p, self.d * self.p)
+        for i in range(self.n):
+            res += torch.kron(YmA_outer[i], xxt_outer[i])
+        return res / self.n
+
+    def getMat_iCnTheta(self, i):
+        a_i = self.A[i].clone().detach()
+        s_i = self.S[i].clone().detach()
+        d_omega = torch.diag(self.Omega)
+        diag_mat_i = torch.diag(1 / a_i + s_i**4 / (1 + s_i**2 * (a_i + d_omega)))
+        Sigma = self.Sigma.clone().detach()
+        return torch.inverse(Sigma + diag_mat_i)
+
+    def getCnTheta(self):
+        xxt_outer = torch.clone(torch.matmul(self.X.unsqueeze(2), self.X.unsqueeze(1)))
+        C_n = torch.zeros(self.d * self.p, self.d * self.p)
+        for i in tqdm(range(self.n)):
+            mat_i = self.getMat_iCnTheta(i)
+            big_mat = torch.zeros(mat_i.shape)
+            big_mat[:] = mat_i[:]
+            xxt_i = xxt_outer[i].clone().detach()
+            C_n += torch.kron(big_mat, xxt_i)
+        return -C_n / self.n
+
+    def getInvSandwich(self):
+        Dn = self.getDnTheta()
+        Cn = self.getCnTheta()
+        return torch.mm(torch.mm(Cn, torch.inverse(Dn)), Cn)
+
+    def getInvFisher(self):
+        vecA = self.A.flatten()
+        # X = torch.zeros((self.n,self.d)).detach().clone()
+        IXt = torch.kron(torch.eye(self.p), self.X).T
+        IX = torch.kron(torch.eye(self.p), self.X)
+        out = torch.multiply(IXt, vecA.unsqueeze(0)) @ (IX)
+        # return torch.inverse(math.sqrt(self.n)*out)
+        return out / self.n
+
+
+def get_cover_from_gaussian(asymptoticGaussianVariational):
+    q1MoinsAlphaSurDeux = 1.96  ## quantile
+    inside = torch.abs(asymptoticGaussianVariational) < q1MoinsAlphaSurDeux
[…]
