+++ viz_rmse.r
@@ -7,11 +7,12 @@ library(cowplot)
 
 
 
-filename = "coverage.csv"
+filename = "rmse.csv"
 df = read.csv(filename)
 df$n = as.factor(df$n)
 df$p = as.factor(df$p)
 df$d = as.factor(df$d)
+
 nb_replicates = max(df$seed) + 1
 df$seed = as.factor(df$seed)
 
@@ -28,13 +29,11 @@ remove_legend <- function(myplot){
 plot_ns_d_p <- function(d,p){
     selected <- (df$d == d) & (df$p == p)
     sub_df <- df[selected,]
-    current_plot <- ggplot(sub_df, aes(x = n, y  = Coverage, fill = method)) + geom_violin(position="dodge", alpha=0.5)
+    current_plot <- ggplot(sub_df, aes(x = n, y  = RMSE, fill = param)) + geom_violin(position="dodge", alpha=0.5)
     # current_plot <- current_plot + stat_summary(fun.data = mean_sdl)
-    current_plot <- current_plot + ylim(0.1,1) + theme_bw() + scale_fill_viridis_d()
-    current_plot <- current_plot + guides(fill=guide_legend(title="Method", nrow = 1, byrow = TRUE))
-    current_plot <- current_plot + geom_hline(yintercept=0.95, color = "red")
+    current_plot <- current_plot + guides(fill=guide_legend(title="", nrow = 1, byrow = TRUE))
     if (p == ps[[1]]){
-        current_plot <- current_plot + ylab(paste("Coverage, d =",as.character(d)))
+        current_plot <- current_plot + ylab(paste("RMSE, d =",as.character(d)))
     }
     else{
         current_plot <- current_plot + ylab("")
@@ -48,11 +47,12 @@ plot_ns_d_p <- function(d,p){
         print(p)
         current_plot <- current_plot + ggtitle(paste("p =",as.character(p)))+
               theme(plot.title = element_text(hjust = 0.5))
+
     }
     return(current_plot)
 }
 
-pdf("plots/coverage.pdf")
+pdf("plots/rmse.pdf")
 plots = list()
 for (d in rev(levels(df$d))){
     for (p in rev(levels(df$p))){
@@ -63,8 +63,8 @@ for (d in rev(levels(df$d))){
     }
 }
 
-title <- paste("Coverage for ", nb_replicates ," replicates on B", sep = "")
+title <- paste("RMSE for ", nb_replicates, " replicates",sep = "")
 
-grid.arrange(grobs = plots, as.table = TRUE, align = c("v"), ncol = 3, bottom = common_legend, common.legend = TRUE,top=textGrob("Coverage for 10 replicates on B",gp=gpar(fontsize=20,font=3)))
+grid.arrange(grobs = plots, as.table = TRUE, align = c("v"), ncol = 3, bottom = common_legend, common.legend = TRUE,top=textGrob(title,gp=gpar(fontsize=20,font=3)))
 
 dev.off()
