### 2024-08-10 #6969 @Bastien-mva
Thank you for your comments and suggestions. We have carefully
considered them and made the necessary amendments to our manuscript and its
accompanying documentation:

Documentation:
- The [README.md](https://github.com/PLN-team/pyPLNmodels/blob/main/README.md) file has been enriched with comprehensive explanations, including the differences between the various models offered by the package. It is also included in the [documentation.](https://bbatardiere.pages.mia.inra.fr/pyplnmodels/)
- We have included a link to a more comprehensive notebook (getting_started.ipynb) in the README.md that provides further explanations.

Manuscript:
- The Summary section has been revised to provide a clear and concise overview of the software's functionality and purpose. Nevertheless, we have made the assumption that readers are familiar with Poisson and Gaussian distributions.
- Figure 1 has been updated to include a caption that explains the legend labels. For Figure 2, we mentioned which curves are related to our software. We have ensured that the models implemented within our package are emphasized.

We hope these revisions enhance the clarity and comprehensibility of our work.

Thank you for your time.
