+++ paper/paper.md
@@ -30,15 +30,15 @@ count data. The Poisson LogNormal(PLN)  [@PLN] and its Principal Component
 Analysis variant PLN-PCA [@PLNPCA] are two-sided latent variable models allowing both
 suitable normalization and analysis of multivariate count data, implemented in this package.
 
-Consider $\mathbf Y$ a count matrix consisting of $n$ rows and $p$ columns. The
-$\mathcal P$ (resp. $\mathcal N$) denotes a Poisson (resp. Normal)
-distribution. It is assumed that each individual $\mathbf Y_i$, that is the $i^{\text{th}}$
+Consider $\mathbf Y$ a count matrix consisting of $n$ rows and $p$ columns.  It is assumed that each individual $\mathbf Y_i$, that is the $i^{\text{th}}$
 row of $\mathbf Y$, is independent of the others and follows a Poisson
 lognormal distribution:
 $$\mathbf Y_{i}\sim \mathcal P(\exp(\mathbf Z_{i})), \quad \mathbf Z_i \sim
 \mathcal N(\mathbf o_i + \mathbf B ^{\top} \mathbf x_i, \boldsymbol{\Sigma}),$$
 where $\mathbf x_i \in \mathbb R^d$ and $\mathbf o_i \in \mathbb R^p$ are
-user-specified covariates and offsets (with default values if not available). The matrix $\mathbf B$ is a $d\times p$
+user-specified covariates and offsets (with default values if not available). The
+$\mathcal P$ (resp. $\mathcal N$) denotes a Poisson (resp. Normal)
+distribution. The matrix $\mathbf B$ is a $d\times p$
 matrix of regression coefficients and $\boldsymbol{\Sigma}$ is a $p\times p$ covariance matrix. The variables $\mathbf Z_i$, known as *latent variables*,
 are not directly observable. However, from a statistical perspective,
 they provide more informative insights compared to the observed variables
