### 2024-08-22 #6969 @LingfengLuo0510
Sorry for the delayed comments

Here are my comments:
(1) Fig2 compares the running time of different packages, can you provide the code used to generate Figure 2.
(2) Fig2 shows the fast computaiton speed of your porposed approach, it would be beneficial to include additional performance metrics.
(3) Line 21: Consider rephrasing to "necessitating dimension reduction techniques suitable for count data."
(4) Line 25: Remove the unnecessary brackets.
(5) Line 34-37: Double-check the punctuation in lines 34-37
(6) When discussing GPU and CPU calculations for time comparison, please specify the exact machine type (e.g., GPU model, CPU cores) used.
(7) Increase the font size in Figure 1 for better readability.
(8) line 47: Verify the accuracy of the citation.
(9)  At the beginning of lines 52 and 56, consider adding a few sentences to introduce the other packages mentioned.
(10) Rephrase the sentences in lines 72-76 for improved clarity and coherence.
(11) Line 80, at the begining of a paragraph, make it more detailed.
(12) Line 80: Choose between "maximizing" and "to maximize" in line 80 based on the intended meaning.
(13) Line 80, "Bound" or "BOund"?
(14) Maintain a consistent citation style throughout the paper, including journal names, volume numbers, and issue numbers.
