+++ viz_consistency.py
@@ -1,4 +1,0 @@
-im
-ns = [250, 500, 1000]
-dims = [20, 40, 50]
-nb_covs = [1, 2, 3]
