+++ benchmark/benchmark.py
@@ -192,7 +192,7 @@ if __name__ == "__main__":
     print("res plnpca", df_results_plnpca)
     df_results_pln.to_csv(f"csv_res_benchmark/python_pln_{DEVICE}.csv")
     df_results_plnpca.to_csv(f"csv_res_benchmark/python_plnpca_{DEVICE}.csv")
-    plot_dict(dict_rt, "Pln", ns)
-    plot_dict(dict_rt, "PlnPCA", ns)
-    plt.savefig(f"paper/illustration.png", format="png")
-    plt.show()
+    # plot_dict(dict_rt, "Pln", ns)
+    # plot_dict(dict_rt, "PlnPCA", ns)
+    # plt.savefig(f"paper/illustration.png", format="png")
+    # plt.show()
