+++ paper/paper.md
@@ -130,18 +130,19 @@ practical computational limits.
 
 
 # Benchmark
-We compare
-\begin{itemize}
-\item PLN and PLN-PCA variants fitted with  \verb|pyPLNmodels| on CPU: \textbf{py-PLN-CPU} and \textbf{py-PLN-PCA-CPU}
-\item PLN and PLN-PCA variants fitted with  \verb|pyPLNmodels| on GPU: \textbf{py-PLN-GPU} and \textbf{py-PLN-PCA-GPU}
-\item PLN and PLN-PCA variants fitted with  \verb|PLNmodels| (on CPU): \textbf{R-PLN} and \textbf{R-PLN-PCA}
-\item \verb|GLLVM| (on CPU): \textbf{GLLVM}
-\end{itemize}
-on the `scMARK` dataset, a benchmark for scRNA data, with
-$n=19998$ cells (samples) and 14059 genes (variables) are available.
-We plot below the running times required to fit such models when the number of variables (i.e.
-genes) grows when $n = 100,1000, 19998$. We used $q =5$ Principal Components when fitting each
-PLN-PCA model and the number of latent variables LV=$2$ for the ```GLLVM``` model.
+
+We conducted a comparison using the following configurations:
+
+- PLN and PLN-PCA models fitted with `pyPLNmodels` on CPU, referred to as **py-PLN-CPU** and **py-PLN-PCA-CPU** respectively.
+- PLN and PLN-PCA models fitted with `pyPLNmodels` on GPU, referred to as **py-PLN-GPU** and **py-PLN-PCA-GPU** respectively.
+- PLN and PLN-PCA models fitted with `PLNmodels` on CPU, referred to as **R-PLN** and **R-PLN-PCA** respectively.
+- The `GLLVM` model fitted on CPU, referred to as **GLLVM**.
+
+These models were tested on the `scMARK` dataset, a benchmark for scRNA data,
+which contains 19,998 cell samples and 14,059 gene variables.
+We plotted the fitting time for these models against an increasing number of
+gene variables, ranging from $5$ to $15000$. Additionally, we varied the number
+of cell samples at $n = 100, 1000, 19998$.
 For each model, the fitting process was halted if the running time exceeded
 10,000 seconds. We were unable to run ```GLLVM``` for $n = 19998$ due to CPU memory
 limitations (64 GB RAM). Similarly, ```py-PLN-PCA-GPU``` could not be run when
@@ -150,9 +151,12 @@ $n=19998$ and $p\geq13000$ as it exceeded the GPU memory capacity (24 GB RAM).
 
 ![Running time analysis on the scMARK benchmark.](figures/plots_benchmark.pdf)
 
-Each package uses variational inference, maximizing an Evidence
-Lower BOund(ELBO) approximating the log-likelihood of the model.
-```GLLVM``` uses an alternate-optimization scheme, fitting alternatively a
+Variational inference aims to approximate the posterior distribution of the
+latent variables by minimizing the divergence between the posterior and a
+variational distribution. Each package uses variational inference
+[@blei2017variational] to maximize an Evidence Lower BOund(ELBO), which serves
+as an approximation to the model's log-likelihood.
+ Each package uses its  ```GLLVM``` uses an alternate-optimization scheme, fitting alternatively a
 Negative Binomial (NB) Generalized Linear Model(GLM) and two penalized NB GLM
 coupled with a fixed-point algorithm, while ```pyPLNmodels``` and
 ```PLNmodels``` uses gradient ascent to maximize the ELBO.
