+++ paper/paper.bib
@@ -2,9 +2,7 @@
     title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
     author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
     booktitle = {Advances in Neural Information Processing Systems 32},
-    pages = {8024--8035},
     year = {2019},
-    publisher = {Curran Associates, Inc.},
     url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
 
@@ -17,8 +15,6 @@
     journal = {Frontiers in Ecology and Evolution},
     year = {2021},
     doi = {10.3389/fevo.2021.588292},
-    url =
-      {https://www.frontiersin.org/articles/10.3389/fevo.2021.588292},
 }
 
 @inproceedings{statsmodels,
@@ -35,10 +31,8 @@
   title        = {scMARK an 'MNIST' like benchmark to evaluate and
                    optimize models for unifying scRNA data},
   year         = {2021},
-  publisher    = {Zenodo},
   version      = {1.0},
   doi          = {10.5281/zenodo.5765804},
-  url          = {https://doi.org/10.5281/zenodo.5765804}
 }
 
 
@@ -47,21 +41,18 @@
     author = {Aitchison, J. and Ho, C. H.},
     title = "{The multivariate Poisson-log normal distribution}",
     journal = {Biometrika},
-    volume = {76},
-    number = {4},
-    pages = {643-653},
     year = {1989},
     month = {12},
     abstract = "{The statistical analysis of multivariate counts has proved difficult because of the lack of a parametric class of distributions supporting a rich enough correlation structure. With increasing availability of powerful computing facilities an obvious candidate for consideration is now the multivariate log normal mixture of independent Poisson distributions, the multivariate Poisson-log normal distribution. The properties of this discrete multivariate distribution are studied and its uses in a variety of applications to multivariate count data are illustrated.}",
     issn = {0006-3444},
     doi = {10.1093/biomet/76.4.643},
-    url = {https://doi.org/10.1093/biomet/76.4.643},
     eprint = {https://academic.oup.com/biomet/article-pdf/76/4/643/706252/76-4-643.pdf},
 }
 @misc{PLNPCA,
       title={Variational inference for probabilistic Poisson PCA},
       author={Julien Chiquet and Mahendra Mariadassou and Stéphane Robin},
       year={2018},
+      journal = {The Annals of Applied Statistics},
       eprint={1703.06633},
       archivePrefix={arXiv},
       doi={10.1214/18-aoas1177},
@@ -75,8 +66,7 @@
       David I. Warton},
     journal = {Methods in Ecology and Evolution},
     year = {2019},
-    volume = {10},
-    pages = {2173-2182},
+    doi = {10.1111/2041-210X.13303}
   }
 
 @misc{PLNzero,
@@ -94,4 +84,12 @@
   author = {Steven G. Johnson},
   year = {2007},
   howpublished = {\url{https://github.com/stevengj/nlopt}}
+}
+
+
+@article{blei2017variational,
+  title={Variational inference: A review for statisticians},
+  author={Blei, David M and Kucukelbir, Alp and McAuliffe, Jon D},
+  journal={Journal of the American statistical Association},
+  year={2017},
 }
