+++ paper/paper.bib
@@ -93,3 +93,12 @@
   journal={Journal of the American statistical Association},
   year={2017},
 }
+
+
+@article{tmb,
+ title={TMB: Automatic Differentiation and Laplace Approximation},
+ doi={10.18637/jss.v070.i05},
+ journal={Journal of Statistical Software},
+ author={Kristensen, Kasper and Nielsen, Anders and Berg, Casper W. and Skaug, Hans and Bell, Bradley M.},
+ year={2016},
+}
