+++ paper/paper.bib
@@ -102,3 +102,11 @@
  author={Kristensen, Kasper and Nielsen, Anders and Berg, Casper W. and Skaug, Hans and Bell, Bradley M.},
  year={2016},
 }
+
+@article{plnmodels_package,
+    title={PLNmodels: Poisson Lognormal Models},
+    author={Chiquet, Julien and Mariadassou, Mahendra and Robin, Stephane},
+    year={2021},
+    url={https://cran.r-project.org/web/packages/PLNmodels/index.html},
+    keywords={package},
+}
