### 2024-08-29 64a4c97
remove the big title in plots.
