### 2024-09-02 #6969 @LingfengLuo0510
@Bastien-mva
My comments have been addressed sufficiently.
Just two more minor things:
(1) Do we need to capitalize the "O" in “BOund”?
(2) Line 115, the citation style is not consistent (no link provided).

Other than that, good to go here. @likeajumprope
