+++ paper/paper.bib
@@ -91,6 +91,7 @@
   title={Variational inference: A review for statisticians},
   author={Blei, David M and Kucukelbir, Alp and McAuliffe, Jon D},
   journal={Journal of the American statistical Association},
+  url={http://dx.doi.org/10.1080/01621459.2017.1285773},
   year={2017},
 }
 
