### 2024-09-03 #6969 @Bastien-mva
I have added the link in the bibliography and corrected the capitalization in 'Bound'.
