### 2024-09-24 bbb1ae7
heatmap of sigma.
