### 2024-10-02 de0e7aa
does not remove samples that are only 0.
