+++ pyPLNmodels/_utils.py
@@ -798,7 +798,8 @@ def _handle_data(
         endog, exog, offsets, offsets_formula, take_log_offsets, add_const
     )
     _check_data_shape(endog, exog, offsets)
-    samples_only_zeros = torch.sum(endog, axis=1) == 0
+    # samples_only_zeros = torch.sum(endog, axis=1) == 0
+    samples_only_zeros = torch.sum(endog, axis=1) < -1
     if torch.sum(samples_only_zeros) > 0.5:
         samples = torch.arange(endog.shape[0])[samples_only_zeros]
         msg = f"The ({len(samples)}) following (index) counts contains only zeros and are removed."
@@ -807,7 +808,7 @@ def _handle_data(
         warnings.warn(msg)
         endog, exog, offsets = _remove_samples(endog, exog, offsets, samples_only_zeros)
         print(f"Now dataset of size {endog.shape}")
-    dim_only_zeros = torch.sum(endog, axis=0) == 0
+    dim_only_zeros = torch.sum(endog, axis=0) < -1
     if torch.sum(dim_only_zeros) > 0.5:
         dims = torch.arange(endog.shape[1])[dim_only_zeros]
         msg = f"The ({len(dims)}) following (index) variables contains only zeros and are removed."
