### 2024-10-09 e3657d5
add fisher variational.
