+++ pyPLNmodels/sampling.py
@@ -198,8 +198,9 @@ def _get_simulation_coef_cov_offsets_coefzi(
         coef = None
     else:
         coef = torch.randn(exog.shape[1], dim)
-        coef /= np.sqrt(coef.shape[0])
-        coef[0, :] += mean_gaussian
+        if nb_cov > 1:
+            coef /= np.sqrt(2)
+        coef += mean_gaussian
     offsets = torch.randint(low=0, high=2, size=(n_samples, dim), dtype=torch.float64)
     torch.random.set_rng_state(prev_state)
     return coef, exog, exog_inflation, offsets, coef_inflation
