### 2024-10-16 ab1d6e2
just right plots.
