### 2024-10-17 af716da
almosth finished visualization of IC.
