### 2024-10-18 a65c1fc
right samplinug parameters, not 1/sqrt(2), but 1
