### 2024-10-26 b56b9be
change simulations. All in one plot.
