### 2024-10-28 db80b63
add fisher information variational.
