### 2024-10-29 f28c526
nicer plot for real data.
