### 2024-10-29 df2fa4c
add alpha =0.5
