### 2024-10-31 a5575ac
get the diagonal elements for the variance instead of a matrix multiplication.
