+++ normality.py
@@ -22,12 +22,16 @@ def rmse(t):
 
 
 nb_max_iter = 800
-nb_seed_param = 1
 nb_seed_count = 100
-ns = [500, 1000, 2000][::-1]
-dims = [25, 50, 100][::-1]
+ns = [1000, 2000, 3000][::-1]
+dims = [50, 100, 200][::-1]
 nb_covs = [1, 2, 3, 4][::-1]
+# ns = [2000][::-1]
+# dims = [100][::-1]
+# nb_covs = [1, 2, 3, 4][::-1]
 mean_gaussian = 2
+
+diag_normalizing = True
 
 N_SAMPLES_KEY = "n_samples"
 P_KEY = "p"
@@ -57,78 +61,74 @@ LIST_KEY = [
 
 def get_each_gaussian(ns, nb_cov, dim):
     tmp_df = pd.DataFrame(columns=LIST_KEY)
-    for seed_param in range(nb_seed_param):
-        sim_param = get_simulation_parameters(
-            n_samples=max(ns),
-            dim=dim,
-            seed=seed_param,
-            nb_cov=nb_cov - 1,
-            add_const=True,
-            mean_gaussian=mean_gaussian,
-        )
-        sim_param._exog = torch.from_numpy(
-            np.random.multinomial(1, [1 / nb_cov] * nb_cov, size=sim_param.n_samples)
-        ).double()
-        for seed_count in range(nb_seed_count):
-            _endog = sample_pln(sim_param, seed=seed_count)
-            _exog = sim_param.exog
-            _offsets = sim_param.offsets
-            true_covariance = sim_param.covariance
-            true_coef = sim_param.coef
-            XB = _exog @ true_coef
-            print(" mean XB", torch.mean(XB))
-            print("min XB", torch.min(XB))
-            for i, n in enumerate(ns):
-                dict_df = {l: [] for l in LIST_KEY}
+    seed_param = nb_cov * dim
+    sim_param = get_simulation_parameters(
+        n_samples=max(ns),
+        dim=dim,
+        seed=seed_param,
+        nb_cov=nb_cov - 1,
+        add_const=True,
+        mean_gaussian=mean_gaussian,
+    )
+    sim_param._exog = torch.from_numpy(
+        np.random.multinomial(1, [1 / nb_cov] * nb_cov, size=sim_param.n_samples)
+    ).double()
+    for seed_count in range(nb_seed_count):
+        _endog = sample_pln(sim_param, seed=seed_count)
+        _exog = sim_param.exog
+        _offsets = sim_param.offsets
+        true_covariance = sim_param.covariance
+        true_coef = sim_param.coef
+        XB = _exog @ true_coef
+        print(" mean XB", torch.mean(XB))
+        print("min XB", torch.min(XB))
+        for i, n in enumerate(ns):
+            dict_df = {l: [] for l in LIST_KEY}
 
-                endog = _endog[:n]
-                exog = _exog[:n]
-                print("endog shape", endog.shape)
-                pln = Pln(endog, exog=exog, add_const=False)
-                pln.fit(nb_max_epoch=nb_max_iter, tol=1e-8)
+            endog = _endog[:n]
+            exog = _exog[:n]
+            print("endog shape", endog.shape)
+            pln = Pln(endog, exog=exog, add_const=False)
+            pln.fit(nb_max_epoch=nb_max_iter, tol=1e-8)
 
-                A = torch.exp(
-                    pln.offsets + pln.latent_mean + 0.5 * pln.latent_sqrt_var**2
-                )
+            A = torch.exp(pln.offsets + pln.latent_mean + 0.5 * pln.latent_sqrt_var**2)
 
-                test = Fisher_Pln(
-                    pln.endog,
-                    A,
[…]
