+++ viz_rmse.r
@@ -8,7 +8,7 @@ library(latex2exp)
 
 
 
-filename = "csvs/res_nb_seed_count100_nb_seed_param_1.csv"
+filename = "csvs/res_nb_seed_count30_diagonal.csv"
 
 df = read.csv(filename)
 df <- subset(df, select = -c(X,Sandwich.based.Information,Variational.Fisher.Information))
