+++ viz_coverage.r
@@ -13,7 +13,7 @@ get_ks_p_value <- function(vector) {
 
 
 # filename = "csvs/res_nb_seed_count100_nb_seed_param_1_diagonal.csv"
-filename = "csvs/res_nb_seed_count30_diagonal.csv"
+filename = "csvs/res_nb_seed_count100_diagonal.csv"
 
 df = read.csv(filename)
 df = subset(df, select = -c(X))
