+++ viz_all.r
@@ -11,7 +11,7 @@ options(error=traceback)
 
 # viz = "dims"
 # perf = "computation"
-viz = "poisson"
+viz = "proba"
 perf = "stat"
 
 pdf(paste("figures/",viz,"_",perf,".pdf",sep=""), height = 10, width = 10)
