+++ normality.py
@@ -22,10 +22,13 @@ def rmse(t):
 
 
 nb_max_iter = 800
-nb_seed_count = 100
-ns = [1000, 2000, 3000][::-1]
-dims = [50, 100, 200][::-1]
-nb_covs = [1, 2, 3, 4][::-1]
+nb_seed_count = 5
+ns = [250, 500, 750, 1000, 1250, 1500, 1750, 2000, 2250, 2500, 2750, 3000][::-1]
+# ns = [1000,2000,3000][::-1]
+dims = [100][::-1]
+# dims = [50, 100, 200][::-1]
+# nb_covs = [1, 2, 3, 4][::-1]
+nb_covs = [2]
 # ns = [2000][::-1]
 # dims = [100][::-1]
 # nb_covs = [1, 2, 3, 4][::-1]
