+++ normality.py
@@ -22,7 +22,7 @@ def rmse(t):
 
 
 nb_max_iter = 800
-nb_seed_count = 5
+nb_seed_count = 100
 ns = [250, 500, 750, 1000, 1250, 1500, 1750, 2000, 2250, 2500, 2750, 3000][::-1]
 # ns = [1000,2000,3000][::-1]
 dims = [100][::-1]
