+++ viz_rmse.r
@@ -6,20 +6,24 @@ library(cowplot)
 library(latex2exp)
 library(viridis)
 
-
+loglog = TRUE
 
 if (loglog == TRUE){
-    filename = "csvs/res_nb_seed_count5_diagonal.csv"
+    filename = "csvs/res_nb_seed_count100_dims_[100]diagonal.csv"
 }else{
     filename = "csvs/res_nb_seed_count100_diagonal.csv"
 }
 df = read.csv(filename)
 df <- subset(df, select = -c(X,Sandwich.based.Information,Variational.Fisher.Information))
+df$theoretical_rmse <- 2.2 / sqrt(df$n_samples)  # Adjust the constant as needed
+# df$theoretical_rmse <- 0.13
 
 # head(df)
 # df <- df[df$RMSE_B < 0.13,]
 # df <- df[df$RMSE_Sigma < 0.13,]
-
+if (loglog == TRUE){
+    # df$n_samples = log(df$n_samples)
+}
 df$n_samples = as.factor(df$n_samples)
 df$p = as.factor(df$p)
 df$nb_cov = as.factor(df$nb_cov)
@@ -42,10 +46,13 @@ label_equal <- function(variable, value){
     return(paste(variable, "=", value))
 }
 
-df = reshape2::melt(df, id.vars = c("n_samples","p", "dim_number", "nb_cov", "seed_count", "seed_param"), value.name = "RMSE")
+df = reshape2::melt(df, id.vars = c("n_samples","p", "dim_number", "nb_cov", "seed_count", "seed_param", "theoretical_rmse"), value.name = "RMSE")
+if (loglog == TRUE){
+    # df$RMSE = log(df$RMSE)
+}
 
 
-df <- df %>% group_by(n_samples, nb_cov ,variable, p, dim_number, seed_count, seed_param) %>% summarise( RMSE = mean(RMSE)) %>% ungroup()
+df <- df %>% group_by(n_samples, nb_cov ,variable, p, dim_number, seed_count, seed_param, theoretical_rmse) %>% summarise( RMSE = mean(RMSE)) %>% ungroup()
 print('head')
 print(head(df))
 
@@ -57,19 +64,29 @@ plot_ns_d_p_facet <- function() {
                     geom_violin(position="dodge", alpha=0.5) +
                     guides(fill=guide_legend(title="", nrow = 1, byrow = TRUE)) +
                     theme_bw() +
+                    # scale_x_log10() +
                     # scale_fill_viridis_d(labels = unname(TeX(c("RMSE$(\\hat{B} - B^{*})$", "RMSE$(\\hat{\\Sigma} - \\Sigma^{*})$")))) +
                     scale_fill_manual(labels = unname(TeX(c("RMSE$(\\hat{B} - B^{*})$", "RMSE$(\\hat{\\Sigma} - \\Sigma^{*})$"))), values = viridis_colors) +
                     facet_grid(nb_cov ~ p, scales = "free", labeller = label_equal) +
                     ylab("RMSE") +
                     xlab(TeX("Number of samples $n$")) +
-                    theme(legend.position="bottom", legend.title = element_blank(), legend.text = element_text(size = rel(4.5)), text = element_text(size = rel(4.5)), strip.text.x = element_text(size = rel(3.5)))+
-                    scale_y_continuous(limits = c(0.025,0.13))
+                    theme(legend.position="bottom", legend.title = element_blank(), legend.text = element_text(size = rel(4.5)), text = element_text(size = rel(4.5)), strip.text.x = element_text(size = rel(3.5)))
+                    # scale_y_continuous(limits = c(0.025,0.13))
[…]
