### 2024-11-14 #6969 @likeajumprope
@Bastien-mva  some small corrections:  line 54: "Notably, It handles" -> it needs to be lowercase; line 33 "functionalities are" -> ":" is missing.
