### 2024-11-14 #6969 @likeajumprope
also, would it be possible to get the quality of Figure 1 to that of Figure 2? Meaning, make the legend and the axis labels and the titles bigger and maybe also the dots? I know that the figure is purely illustrative, but the dots of other_myeloids and T_cells_CD4+ are not distinguishable.
Figure 2 is beautiful.
