+++ paper/paper.bib
@@ -75,7 +75,8 @@
       year={2024},
       eprint={2405.14711},
       archivePrefix={arXiv},
-      primaryClass={stat.ME}
+      primaryClass={stat.ME},
+      doi = {10.21203/rs.3.rs-4637786/v1}
 }
 
 
@@ -92,6 +93,7 @@
   author={Blei, David M and Kucukelbir, Alp and McAuliffe, Jon D},
   journal={Journal of the American statistical Association},
   url={http://dx.doi.org/10.1080/01621459.2017.1285773},
+  doi = {10.1080/01621459.2017.1285773},
   year={2017},
 }
 
@@ -109,5 +111,6 @@
     author={Chiquet, Julien and Mariadassou, Mahendra and Robin, Stephane},
     year={2021},
     url={https://cran.r-project.org/web/packages/PLNmodels/index.html},
+    doi = {10.32614/cran.package.plnmodels},
     keywords={package},
 }
