### 2024-11-15 080baf4
add bigger points and remarks of JOSS.
