### 2024-11-15 85689cf
bigger points in the fig and the legend.
