### 2024-11-15 #6969 @Bastien-mva
Thank you for your feedback, @likeajumprope .

I have addressed your suggestions as follows:

- Updated the DOIs.
- Corrected the two typos.
- Enhanced Figure 1 for readability by increasing point sizes, enlarging titles and axis labels, and adjusting colors to better distinguish between other_myeloids and T cells CD4+.
