### 2024-11-18 #6969 @Bastien-mva
I noticed colors were not right between the two plots of the first figure. This is now fixed.
