+++ viz_all_these.r
@@ -1,0 +1,200 @@
+library(ggplot2)
+library(cowplot)
+library(gridExtra)
+library(dplyr)
+library(hash)
+library(latex2exp)
+library(grid)
+library(glue)
+
+options(error=traceback)
+
+viz = "proba"
+perf = "stat"
+
+pdf(paste("figures/",viz,"_",perf,"_column_dependent.pdf",sep=""), height = 10, width = 30)
+
+get_name_computation <- function(viz,formula){
+    if (viz == "dims"){
+        return(paste(viz,formula,"not_n_or_p_500.csv", sep = "_"))
+    }
+    else{
+        return(paste(viz,formula,"not_n_or_p_150.csv", sep = "_"))
+    }
+}
+
+name_doss_1 = paste(viz,"_viz_global_right_simu_multin.csv", sep = "")
+name_doss_2 = paste(viz,"_viz_column-wise_right_simu_multin.csv", sep = "")
+name_doss_3 = paste(viz,"_viz_row-wise_right_simu_multin.csv", sep = "")
+name_dosses = c(name_doss_2)
+
+h = hash()
+h[["ELBO"]] = "ELBO"
+h[["Reconstruction_error"]] =TeX('RMSE $Y$')
+h[["RMSE_OMEGA"]] = TeX('RMSE $\\Omega$')
+h[["RMSE_SIGMA"]] = TeX('RMSE $\\Sigma$')
+h[["RMSE_B"]] =TeX('RMSE $B$')
+h[["RMSE_PI"]]=  TeX('RMSE $\\pi$')
+h[["RMSE_B0"]] =TeX('RMSE $B^0$')
+h[["TIME"]] = "Time"
+h[["NBITER"]] = "Number of iterations"
+
+g_legend<-function(a.gplot){
+      tmp <- ggplot_gtable(ggplot_build(a.gplot))
+  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
+    legend <- tmp$grobs[[leg]]
+    return(legend)}
+scaleFUN <- function(x) sprintf("%.4s", x)
+
+get_df = function(namedoss, perf, viz){
+    columns = c("model_name")
+    if (perf == "stat"){
+        if (viz == "samples" || viz == "dims"){
+            columns = c(columns,"xscale")
+        }
+        else{
+            columns = c(columns,"moyenne")
+        }
+        columns = c(columns,"RMSE_SIGMA","RMSE_PI","RMSE_B")
+    }
+    if (perf == "computation"){
+        if (viz == "poisson" || viz == "proba"){
+            columns = c(columns, "moyenne")
+        }
+        else{
+            columns = c(columns,"xscale")
+        }
+        columns = c(columns,"TIME","Reconstruction_error")
+    }
+    if (perf == "annexe"){
+        columns = c(columns,"moyenne","RMSE_OMEGA","RMSE_PI","RMSE_B","RMSE_B0")
+    }
+    df = subset(read.csv(paste('csv_data/',namedoss, sep = "")), select = -X)
+    for (column in colnames(df)){
+        df[df == 666] = NA
+        }
+    df[df$model_name == "Pln",]$model_name <- "PLN"
+    if (viz == "samples" || viz == "dims"){
+        df[df$model_name == "fair_Pln",]$model_name <- "Oracle PLN"
+    }
+    else{
+        df[df$model_name == "Fair Pln",]$model_name <- "Oracle PLN"
+    }
+    df <- df[df$model_name %in% c("Enhanced Analytic", "Enhanced", "Standard", "Standard Analytic"),]
+    df[df$model_name == "Enhanced Analytic",]$model_name <- "Amélioré (Analytique)"
[…]
