### 2024-11-18 #6969 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
