+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @incollection{Pytorch,
-    title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+    title = {{P}y{T}orch: An Imperative Style, High-Performance Deep Learning Library},
     author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
     booktitle = {Advances in Neural Information Processing Systems 32},
     year = {2019},
@@ -10,7 +10,7 @@
 @Article{PLNmodels,
     author = {Julien Chiquet and Mahendra Mariadassou and Stéphane
       Robin},
-    title = {The Poisson-lognormal model as a versatile framework for
+    title = {The {P}oisson-lognormal model as a versatile framework for
       the joint analysis of species abundances},
     journal = {Frontiers in Ecology and Evolution},
     year = {2021},
@@ -28,7 +28,7 @@
 
 @dataset{scMark,
   author       = {Javier Diaz-Mejia},
-  title        = {scMARK an 'MNIST' like benchmark to evaluate and
+  title        = {sc{MARK} an 'MNIST' like benchmark to evaluate and
                    optimize models for unifying scRNA data},
   year         = {2021},
   version      = {1.0},
@@ -39,7 +39,7 @@
 
 @article{PLN,
     author = {Aitchison, J. and Ho, C. H.},
-    title = "{The multivariate Poisson-log normal distribution}",
+    title = "{The multivariate {P}oisson-log normal distribution}",
     journal = {Biometrika},
     year = {1989},
     month = {12},
@@ -49,7 +49,7 @@
     eprint = {https://academic.oup.com/biomet/article-pdf/76/4/643/706252/76-4-643.pdf},
 }
 @misc{PLNPCA,
-      title={Variational inference for probabilistic Poisson PCA},
+      title={Variational inference for probabilistic {P}oisson PCA},
       author={Julien Chiquet and Mahendra Mariadassou and Stéphane Robin},
       year={2018},
       journal = {The Annals of Applied Statistics},
@@ -70,7 +70,7 @@
   }
 
 @misc{PLNzero,
-      title={Zero-inflation in the Multivariate Poisson Lognormal Family},
+      title={Zero-inflation in the Multivariate {P}oisson Lognormal Family},
       author={Bastien Batardière and Julien Chiquet and François Gindraud and Mahendra Mariadassou},
       year={2024},
       eprint={2405.14711},
