### 2024-12-04 #6969 @Bastien-mva
> 👋 @Bastien-mva - I just need you to address the following items before I can accept this for publication:
>
> In the archive:
>
>     * [ x]  I don't see a current version set from a GitHub release of the software.  Your latest release on GitHub is `v0.0.1` and the version listed in your Zenodo archive is `v1`.  Please create a release of the current codebase in GitHub and ensure that that release number matches what is in your Zenodo archive.  Once you do so, please list the version number here in this thread (cc. @likeajumprope).
>
>
> In the paper:
>
>     * [ x]  LINE 8:  Missing space between "LogNormal" and "(PLN)".  Also, I believe the non-acronym part is normally written "Poisson-log normal" in the original 1989 work.
>
>     * [ x]  LINE 15:  For clarity, use "respectively" in both places instead of "resp."
>
>     * [ x]  LINE 48:  The "t" in "Pytorch" should be capitalized
>
>     * [ x]  LINE 52:  "The statsmodels (Seabold & Perktold, 2010) is a Python library providing..." should be written "The statsmodels (Seabold & Perktold, 2010) Python packages provides..."
>
>     * [ x]  LINES 54-55:  Should be written: "...through the Generalized Linear Model class implementations PoissonBayesMixedGLM and BinomialBayesMixedGLM."
>
>     * [ x]  LINES 89-90:  Missing space between “Evidence Lower Bound” and “(ELBO)”
>
>     * [ x]  LINES 92-93:  "...all the methods uses gradient..." should be "...all of the methods use gradient..."
>
>     * [ x]  LINE 95:  Missing space between "nlopt" and "(Johnson, 2007)"
>
>     * [ x]  LINE 96:  The "t" in "Pytorch" should be capitalized
>
>     * [ x]  Please ensure all capitalization is correctly maintained in all of your references (e.g., LINE 113, the "p" in "poisson" should be capitalized).  You can do this by placing the characters you wish to maintain the formatting of in curly brackets in your bib file.
>
>
> Once you have made these changes, please notify me here. Thanks!

Thanks for the remarks. I have just adressed them. For the version, I have made a release version for github named v1. I am a little bit confused between zenodo versions, package versions and release namse of GitHub. The release I have done may not be adequate.
