### 2024-12-05 #6969 @Bastien-mva
Thanks everyone for all your work ! It is really great to have such an open-source journal. Thanks again !
